+++ src/modules/broadcast/interfaces/Messagable.ts
